Agregando el método Get obtener un topico

+++ src/main/java/com/alura/ForoAlura/domain/topicos/TopicoRepositorio.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface TopicoRepositorio extends JpaRepository<Topico, Integer> {
+public interface TopicoRepositorio extends JpaRepository<Topico, Long> {
     boolean existsByTituloAndMensaje(String titulo, String mensaje);
 }
